add Component stuff
* add Text to button

* Optimise Font

* add Focus and basic text input in TextBox

* add UTF8 support for text input

* add formatted logging boi

* finalize TextBox

* add Scheduler and coroutines

* make buttons scriptable

* Component stores pointer to layer

* add component ID

* fix Scheduler

* Integrate Camera in the component system

+++ zoe/src/zoe/core/Core.h
@@ -42,5 +42,5 @@
 #endif
 
 #define PI 3.141592f
-#define BIT(X) (1 << X)
+#define BIT(X) (1u << (uint32_t)X)
